requirements.txt
torch==2.1.0
torchvision==0.15.0
matplotlib==3.7.1
kagglehub==1.0.0
numpy==1.24.2
scikit-learn==1.2.2
Pillow==9.5.0
opencv-python==4.7.0.72
streamlit
flask